dev_commit: implemented switching between snapshots & fixes

+++ lua/nvchad/updater/snap_delete.lua
@@ -7,66 +7,80 @@ local function snap_delete()
    local echo = utils.echo
    local current_branch_name = git.get_current_branch_name()
 
+   -- BETA DISCLAIMER: This feature is still in BETA. If you encounter any issues contact
+   -- @LeonHeidelbach on GitHub by creating an issue in the main NvChad repo.
+   echo(prompts.beta_disclaimer)
+   misc.print_padding("\n", 1)
+
    local snapshot_list = git.get_branch_list(function(branch)
       return branch:match(defaults.snaps.base_snap_branch_name)
    end)
 
    if #snapshot_list == 0 then
-      echo({ { "No snapshots found." } })
+      echo(prompts.no_snapshots_found)
       return
    end
 
    local snapshot_list_str = {}
 
-   vim.list_extend(snapshot_list_str, { { "Select the snapshot(s) that you would like to delete:\n", "WarningMsg" } })
+   vim.list_extend(snapshot_list_str, prompts.select_snapshot_to_delete)
 
    for i, snapshot in ipairs(snapshot_list) do
-      if snapshot:match(current_branch_name) then snapshot_list[i] = current_branch_name end
-      vim.list_extend(snapshot_list_str, { { "    [" .. tostring(i) .. "] " }, { snapshot, "Title" },
-         { (current_branch_name == snapshot_list[i] and " (currently in use)" or "") }, { "\n" } })
+      if snapshot:match("\\* " .. current_branch_name .. "$") then
+         snapshot_list[i] = current_branch_name
+      end
+
+      vim.list_extend(snapshot_list_str, { { "    [" .. tostring(i) .. "] " }, { snapshot,
+         "Title" }, { (current_branch_name == snapshot_list[i] and " (currently in use)" or "") },
+         { "\n" } })
    end
+
+   vim.list_extend(snapshot_list_str, prompts.select_snapshot_to_delete_enter_index)
 
    echo(snapshot_list_str)
 
-   echo({ { "\nEnter the indices of the snapshots separated by space or [a]bort [<number_list>/A]:",
-      "WarningMsg" } })
-
-   local selection = string.lower(vim.fn.input("-> "))
+   local selection = vim.trim(string.lower(vim.fn.input("-> ")))
 
    misc.print_padding("\n", 2)
 
-   if selection == "a" then
-      echo({ { "Aborted." } })
+   if selection == "c" then
+      echo(misc.list_text_replace(prompts.cancelled_action, "<ACTION>", "NvChadSnapshotDelete"))
       return
    end
 
    local selection_list = vim.fn.split(selection, " ")
 
    if #selection_list == 0 then
-      echo({ { "No snapshots selected." } })
+      echo(prompts.no_snapshots_selected)
       return
    end
 
    local processed = {}
-   local invalid_selection_list = { { "Invalid inputs: ", "WarningMsg" } }
-   local valid_selection_list = { { "The following snapshots have been successfully deleted: ", "WarningMsg" } }
+   local invalid_selection_list = vim.deepcopy(prompts.invalid_inputs)
+   local valid_selection_list = vim.deepcopy(prompts.snapshot_successfully_deleted)
 
    for _, input_str in ipairs(selection_list) do
       local number = tonumber(input_str)
 
-      if not vim.tbl_contains(processed, input_str) then
+      if input_str ~= "" and not vim.tbl_contains(processed, input_str) then
          table.insert(processed, input_str)
 
          if type(number) ~= "number" or number < 1 or number > #snapshot_list or
              current_branch_name == snapshot_list[number] then
 
             if current_branch_name == snapshot_list[number] then
-               echo({ { "Error: You cannot delete a snapshot that is currently in use!\n\n", "ErrorMsg" } })
+               echo(prompts.cannot_delete_current_snapshot)
             end
             vim.list_extend(invalid_selection_list, { { (#invalid_selection_list > 1 and ", "
                 or "") .. input_str } })
          else
+
+            -- delete the snapshot
             if git.delete_branch(snapshot_list[number]) then
+
+               -- delete the corresponding "PackerSnapshot"
+               vim.cmd("PackerSnapshotDelete " .. snapshot_list[number])
+
                vim.list_extend(valid_selection_list,
                   { { (#valid_selection_list > 1 and ", " or "") .. "[" .. input_str .. "] " ..
                       snapshot_list[number] } })
@@ -83,7 +97,7 @@ local function snap_delete()
    if #valid_selection_list > 1 then
       echo(valid_selection_list)
    else
-      echo({ { "No snapshots were deleted!", "WarningMsg" } })
+      echo(prompts.no_snapshots_deleted)
       return
    end
 end
